add comments

+++ Practice 3/2.cpp
@@ -3,7 +3,8 @@
 
 using namespace std;
 
-int cubeRoot(double a, double temp = 1.0, double epsilon = 1e-9)
+// Функция для вычисления кубического корня из числа иттеративным методом
+double cubeRoot(double a, double temp = 1.0, double epsilon = 1e-9)
 {
     double x = (2.0 * temp + a / (temp * temp)) / 3.0;
     if (abs(x - temp) < epsilon)
@@ -23,3 +24,5 @@ int main()
     cout << "Кубический корень из " << a << " по встроенной функции pow() равен \t" << pow(a, 1.0/3) << endl;
     return 0;
 }
+
+// Отличие между написанной функцией и встроенной функцией pow() в ограниченности функции cubeRoot значением epsilon
